requirements.txt
django
djangorestframework
opencv-python
numpy
ultralytics
inference-sdk
google-cloud-vision
torch 